+++ Gamp-SCPeriop.Shared/Entity/Model/ModuleComponentCreateDTO.cs
@@ -13,7 +13,7 @@ namespace GamP_SCPeriop.Shared.Entity.Model
         public string? PdfFilePath { get; set; }
         public ModuleStage Stage { get; set; }
         public int? ParentComponentId { get; set; }
-        public int Weight { get; set; }
+        public float Weight { get; set; }
         public bool IsFromTemplate { get; set; }
     }
 }
